Added Square and SquareRoot functions to calculator . . . square() will set operator to multiplication and pass the operator as both parameters root() will set the operator to 'r' and pass the operator and null
findResult() now has root functionality. The operand is passed through Math.sqrt() and the resulting double is casted to an integer. The resulting binary string is returned.

Added corresponding functions to CalculatorTest

************************************************************************
Other Changes
.
.
.
Restricted access to Calculator Class variables. Values must now be set and retrieved using the getter and setter functions.

String binaryString was replaced with a StringBuilder. This prevents the program from having to create a new string every time a digit is appended to the string.